[#44] refactor : enterPage 로 컴포넌트 이동
